implement saving a gig

+++ ViewModels/GigFormViewModel.cs
@@ -4,11 +4,14 @@ using System.Collections.Generic;
 namespace GigHub.ViewModels
 {
     public class GigFormViewModel
-    {
+    {       
+
         public string Venue { get; set; }
         public string Date { get; set; }
         public string Time { get; set; }
-        public int Genre { get; set; }
+        public byte Genre { get; set; }
         public IEnumerable<Genre> Genres { get; set; }
+
+
     }
 }
